requirements.txt
# Telegram Bot
python-telegram-bot>=20.0
python-telegram-bot[job-queue]>=20.0

# Database
mongoengine>=0.24.0
pymongo>=4.0.0

# AI/ML
openai>=1.0.0

# Environment and Configuration
python-dotenv>=0.19.0

# Date/Time Handling
pytz>=2021.3

# Utilities
requests>=2.26.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0

# Development Tools
black>=22.0.0  # Code formatting
flake8>=4.0.0  # Code linting
mypy>=0.910    # Type checking
